+++ packages/model/src/dto/chat/index.ts
@@ -1,5 +1,4 @@
 export * from './send-command.dto';
 export * from './send-question.dto';
-export * from './send-question-response.dto';
 export * from './message.dto';
 export * from './pagination.dto'