+++ science/app/typedefs.py
@@ -99,6 +99,12 @@ class ChatCompletionResponse(TypedDict):
     choices: List[ChatGptChoice]
 
 
+class ChatGptStreamChoice(TypedDict):
+    delta: MessageDict  # no function_ids or webhook_ids
+    finish_reason: Literal['stop', 'length', 'content_filter', None]
+    index: int
+
+
 class StatsDict(TypedDict, total=False):
     prompt: str
     match_count: int