+++ src/chat/chat.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { ChatText } from '@poly/model';
 import { PrismaService } from 'prisma/prisma.service';
 import { AiService } from 'ai/ai.service';
+import { Observable } from 'rxjs';
 
 @Injectable()
 export class ChatService {
@@ -9,16 +9,8 @@ export class ChatService {
 
   constructor(private readonly aiService: AiService, private readonly prisma: PrismaService) {}
 
-  public async getMessageResponseTexts(environmentId: string, userId: string, message: string): Promise<ChatText[]> {
-    const { answer, stats } = await this.aiService.getFunctionCompletion(environmentId, userId, message);
-
-    return [
-      {
-        type: 'markdown',
-        value: answer,
-        stats,
-      },
-    ];
+  public sendQuestion(environmentId: string, userId: string, message: string): Observable<string> {
+    return this.aiService.getFunctionCompletion(environmentId, userId, message);
   }
 
   async processCommand(environmentId: string, userId: string, command: string) {